[update] 메인 수정

+++ src/main/resources/static/js/user/userHeart.js
@@ -46,6 +46,11 @@ function deleteHearts(event) {
         items.push(checkbox.value);
     });
 
+    if(items == null || items.length == 0) {
+        alert("상품을 선택해 주세요!");
+        return false;
+    }
+
     $.ajax({
         type: "DELETE",
         url: '/user/heart/',
@@ -64,15 +69,20 @@ function deleteHearts(event) {
 function deleteAll(event) {
     event.preventDefault();
 
-    if(!confirm("좋아요 목록을 전부 삭제하시겠습니까?")) {
-        return;
-    }
-
     let items = [];
     const checkboxes = document.querySelectorAll('input[name="agreeCheck"]');
     checkboxes.forEach((checkbox) => {
         items.push(checkbox.value);
     });
+
+    if(items == null || items.length == 0) {
+        alert("상품을 선택해 주세요!");
+        return false;
+    }
+
+    if(!confirm("좋아요 목록을 전부 삭제하시겠습니까?")) {
+        return;
+    }
 
     $.ajax({
         type: "DELETE",
